Gemfile: Add database Gems to separate groups
[GitLab](https://github.com/gitlabhq/gitlabhq/blob/master/Gemfile#L12),
for example, adds the database Gems mysql2 and pg to separate groups,
allowing to exclude the installation of the unneeded Gems.

With the command below, the user can avoid installing the the unneeded
Gems and their dependencies, especially as both Gems, mysql2 and pg,
need installed header files for successful installation.

	bundle install --deployment --without development test mysql aws