cms: Update Devise mailer views for Devise 3.1
Since upgrading from Devise 2.2.4 to Devise 3.4.1 in commit 97473a9f
(* update to rails 4.1.7) the password reset is broken. The user
receives an email, but the included token is invalid, so trying to
change the password fails.

Here is the description of the change in the blog post *Devise 3.1: Now
with more secure defaults* [1].

----------------- 8< ------------------------- >8 -----------------
Store digested tokens in the database

In previous versions, Devise stored the tokens for confirmation, reset
password and unlock directly in the database. This meant that somebody
with read access to the database could use such tokens to sign in as
someone else by, for example, resetting their password.

In Devise 3.1, we store an encrypted token in the database and the
actual token is sent only via e-mail to the user. This means that:

* Devise now requires a `config.secret_key` configuration. As soon as
you boot your application under Devise 3.1, you will get an error with
information about how to proceed;
* Every time the user asks a token to be resent, a new token will be
generated;
* The Devise mailer now receives one extra token argument on each
method. If you have customized the Devise mailer, you will have to
update it.  All mailers views also need to be updated to use @token, as
shown here, instead of getting the token directly from the resource;
* Any previously stored token in the database will no longer work unless
you set config.allow_insecure_token_lookup = true in your Devise
initializer. We recomend users upgrading to set this option on
production only for a couple days, allowing users that just requested a
token to get their job done.

Thanks to Stephen Touset for reporting this issue and working with us on
a solution.
----------------- 8< ------------------------- >8 -----------------

So do as they say and use `@token` instead of
`@resource.reset_password_token` in the Devise mailer view.

[1] http://blog.plataformatec.com.br/2013/08/devise-3-1-now-with-more-secure-defaults/
[2] https://github.com/plataformatec/devise/blob/2a8d0f9beeb31cd2287094c5dcf843d0bd069eb8/app/views/devise/mailer/reset_password_instructions.html.erb#L5