initializers/session_store: Use `CacheStore`
`MemCacheStore` is a legacy implemention, it also forces Dalli down
one’s throat, requiring a running Memcached server, which is not always
available, especially during development.

From `guides/source/action_controller_overview.md`:

```
Session
-------

Your application has a session for each user in which you can store
small amounts of data that will be persisted between requests. The
session is only available in the controller and the view and can use one
of a number of different storage mechanisms:

* `ActionDispatch::Session::CookieStore` - Stores everything on the
  client.
* `ActionDispatch::Session::CacheStore` - Stores the data in the Rails
  cache.
* `ActionDispatch::Session::ActiveRecordStore` - Stores the data in a
  database using Active Record. (require `activerecord-session_store`
  gem).
* `ActionDispatch::Session::MemCacheStore` - Stores the data in a
  memcached cluster (this is a legacy implementation; consider using
  CacheStore instead).
```

Normally the cookie-based store is recommended as it’s the fastest [1],
but it is changed in commit 735c8582 (* workflow-updates) to use the
Memcached cluster without any explanation.

So, use the generic Rails cache, which will use the cache store
configured individually for each environment.

The README of the Gem Dalli also gives that example [2].

```
Rails.application.config.session_store ActionDispatch::Session::CacheStore, :expire_after => 20.minutes
```

[1] http://api.rubyonrails.org/classes/ActionDispatch/Session/CookieStore.html
[2] https://github.com/petergoldstein/dalli#usage-with-rails-3x-and-4x