Remove `PAPERCLIP_COMMAND_PATH`
Currently, uploading a file in Debian Stretch/testing in development
mode, it fails, because the utility `convert` is in `/usr/bin`, which is
in the PATH.

There is no need to force the location of the utility. If it’s stored in
`/usr/local/bin`, then that directory should be present in the `PATH`
variable.

So remove the workaround.

```
$ env | grep '^PATH='
PATH=/usr/local/bin:/usr/bin:/bin:/usr/local/games:/usr/games
```

Fixes: #59