Add .travis.yml
The .travis.yml file was copied from the Travis CI Getting started guide
which can be found in this link:https://docs.travis-ci.com/user/getting-started/
or on the Travis CI homepage.

The file is added to initialize Travis CI.

Travis CI is a continuous integration service used to test and build
software. It enables open source projects to be tested.
https://en.wikipedia.org/wiki/Travis_CI
Currently the tests are failing due to a false configuration
in the -travis.yml file. This will be fixed in future commits.